feach: add init.js
