+++ virtual-keyboard/js/index.js
@@ -1,20 +1,12 @@
 
 
 import KeyClass, { createKey } from './keyClass.js'
+import { init } from './init.js'
 
-const BODY = document.querySelector('#body');
-
-const TEXTAREA=document.createElement('textarea');
-console.log('**',TEXTAREA)
-TEXTAREA.id='textarea'
-
-const KEYBOARD=document.createElement('div');
-KEYBOARD.id='keyboard'
+init()
 
 
 
-BODY.append(TEXTAREA)
-BODY.append(KEYBOARD)
-console.log('**',BODY)
+
 const k = new KeyClass('E', 100)
 createKey(k)
